+++ app/client/components/hooks/useWordsForSkull.ts
@@ -88,9 +88,9 @@ function useWordsForSkull({ currentSkull, wordsData }: PropType) {
   }, [currentSkull]);
 
   //Displays list of all answers for debugging purposes
-  // useEffect(() => {
-  //   wordsForSkull.length > 0 && console.log(wordsForSkull);
-  // }, [wordsForSkull]);
+  useEffect(() => {
+    wordsForSkull.length > 0 && console.log(wordsForSkull);
+  }, [wordsForSkull]);
 
   return { wordsForSkull, wordsList, dispWordHistory, setDispWordHistory };
 }
